o  deprecated getObjectByID() methods that use 'int' as an Id. Use version with 'long' o  related updates to unit tests.

+++ plexus-components/trunk/plexus-jdo2/src/test/java/org/codehaus/plexus/jdo/Basic.java
@@ -26,7 +26,7 @@ public class Basic implements java.io.Serializable {
     /**
      * Field id
      */
-    private int id = 0;
+    private long id = 0;
 
     /**
      * Field name
@@ -72,7 +72,7 @@ public class Basic implements java.io.Serializable {
     /**
      * Get id
      */
-    public int getId()
+    public long getId()
     {
         return this.id;
     }  
@@ -110,7 +110,7 @@ public class Basic implements java.io.Serializable {
      * 
      * @param id
      */
-    public void setId(int id)
+    public void setId(long id)
     {
         this.id = id;
     }
